split-common-into-common-checks-io
