+++ cooltools/__init__.py
@@ -14,6 +14,7 @@ import logging
 __version__ = "0.5.0"
 
 from . import lib
+
 from .lib import (
     numutils,
     download_data,
