+++ cooltools/cli/eigs_cis.py
@@ -3,7 +3,8 @@ import numpy as np
 import cooler
 import bioframe
 from ..api import eigdecomp
-from ..lib.common import make_cooler_view, read_viewframe
+from ..lib.common import make_cooler_view
+from ..lib.io import read_viewframe
 
 import click
 from .util import TabularFilePath, sniff_for_header
@@ -192,12 +193,10 @@ def eigs_cis(
         clip_percentile=99.9,
         sort_metric=None,
     )
-    
+
     # Output
     eigvals.to_csv(out_prefix + ".cis" + ".lam.txt", sep="\t", index=False)
-    eigvec_table.to_csv(
-        out_prefix + ".cis" + ".vecs.tsv", sep="\t", index=False
-    )
+    eigvec_table.to_csv(out_prefix + ".cis" + ".vecs.tsv", sep="\t", index=False)
     if bigwig:
         bioframe.to_bigwig(
             eigvec_table,
