+++ cooltools/cli/expected_trans.py
@@ -4,10 +4,13 @@ from itertools import combinations
 import cooler
 import bioframe
 from .. import api
-from ..lib import common
+from ..lib.common import make_cooler_view
+from ..lib.io import read_viewframe
+
 
 import click
 from . import cli
+
 
 @cli.command()
 @click.argument("cool_path", metavar="COOL_PATH", type=str, nargs=1)
@@ -78,10 +81,10 @@ def expected_trans(
 
     if view is None:
         # full chromosome case
-        view_df = common.make_cooler_view(clr)
+        view_df = make_cooler_view(clr)
     else:
         # Read view_df dataframe, and verify against cooler
-        view_df = common.read_viewframe(view, clr, check_sorting=True)
+        view_df = read_viewframe(view, clr, check_sorting=True)
 
     result = api.expected.expected_trans(
         clr,
