+++ cooltools/cli/pileup.py
@@ -4,7 +4,9 @@ import cooler
 import bioframe
 
 from .. import api
-from ..lib import common
+
+from ..lib.common import make_cooler_view
+from ..lib.io import read_viewframe, read_expected
 
 import click
 from functools import partial
@@ -129,7 +131,7 @@ def pileup(
     """
 
     clr = cooler.Cooler(cool_path)
-    cooler_view_df = common.make_cooler_view(clr)
+    cooler_view_df = make_cooler_view(clr)
 
     #### Read the features:
     buf, names = sniff_for_header(features)
@@ -180,7 +182,7 @@ def pileup(
         view_df = cooler_view_df
     else:
         # Read view_df dataframe, and verify against cooler
-        view_df = common.read_viewframe(view, clr, check_sorting=True)
+        view_df = read_viewframe(view, clr, check_sorting=True)
 
     # make sure feature are compatible with the view_df
     if not bioframe.is_contained(features_df, view_df):
@@ -194,7 +196,7 @@ def pileup(
         expected_value_cols = [
             expected_value_col,
         ]
-        expected = common.read_expected(
+        expected = read_expected(
             expected_path,
             contact_type="cis",
             expected_value_cols=expected_value_cols,
