+++ cooltools/lib/__init__.py
@@ -1,1 +1,3 @@
 from .common import *
+from .io import *
+from .checks import *
